added support for building Windows fluent-bit artifacts and aws-for-fluent-bit Windows image
In order to add support for building Windows fluent-bit artifacts, we have created a Powershell script which can be invoked from a CodeBuild project. The generated zip archive has the required binaries and configs.

For building the Windows aws-for-fluent-bit image, we have added the dockerfile and other scripts which can be used for building release specific Windows images.