adds the publish image script for Windows images
For Windows, in order to publish the images, we would be creating EC2 instances for the underlying platform. We would then be using this script to pull the image from source and push the same to target registry/repository.

This script supports the following scenarios-
- Pull the image from regional ECR repository and push the image to another regional repository.
- Pull the image from regional ECR repository and push the image to public ECR repository.
- Pull the image from regional ECR repository and push the image to dockerhub repository.
- Pull the image from public ECR repository and push the image to regional ECR repository.
- Pull the image from public ECR repository and push the image to another public ECR repository.
- Pull the image from public ECR repository and push the image to dockerhub repository.

In our workflows, we make use of the first 4 scenarios.