added the scripts for verifying images and determining if sync required
As part of our publishing workflow, we are pushing the images as well as the manifest into the target repository of the target registry. The next step would then verify if the images are infact present in the target repository.

Additionally, we have a sync task workflow, wherein we need to determine if we want to sync the images in regional ECR repository with the ones present in public ECR. Therefore, we compare the manifests of the same and then export a variable in CodeBuild project. This would be read by the subsequent stages in Step Function to determine if we do need to publish images into the regional ECR repository.