added the scripts for publishing and verifying SSM Parameters
For AWS-for-fluent-bit images, we publish SSM Parameters in each regional store. The normal CodePipeline runs in the leader region of standard partition. As part of that workflow, we publish the SSM Parameters based on new images for each of the classic region in standard partition.

As part of sync-task workflow, we publish the SSM Parameter in each opt-in region as well as China and Gov cloud partition. This sync is based on the updated images in Public ECR repository.

As part of this change, we are adding the following-
- Script to publish and sync SSM parameter in each region
- Script to verify the SSM parameter in each region
- Buildspecs used by CodeBuild projects