windows: fixed bug where the version was not populated properly
In a dockerfile, any ARG declared before FROM is not valid in the build context. Additionally, on Windows, the ARG inside build context needs to be referenced using "$env:". Due to these issues, the AWS_FOR_FLUENT_BIT version was not getting populated properly in the images. This change rectifies the same.