build: separate plugin Dockerfiles for linux and windows
We got this error:
[proxy] error opening plugin /fluent-bit/firehose.so: '/lib64/libc.so.6: version `GLIBC_2.32' not found (required by /fluent-bit/firehose.so)'

This happens because the previous Dockerfile.plugins tried to build both windows and linux plugins in Ubuntu. However, the final Fluent Bit
build for linux happens in Amazon Linux. In seems that AL and Ubuntu use different versions of glibc, and for the cgo shared objects to
work, both them and Fluent Bit need to be built with the same glibc version.

There may be other ways of fixing this that keep the same Dockerfile for both windows and linux, but this fix was quick and simple
and unblocks the pipeline for a CVE release. It also does not add significant new maintenance overhead.

Signed-off-by: Wesley Pettit <wppttt@amazon.com>