windows: consolidate configurable versions in the config file
Presently, we have OpenSSL and FlexBison versions hardcoded in the build scripts itself. This means that changing the same requires a change to the build script. Additionally, the versions are then scattered across all the build files making it harder to understand.

Therefore, as part of this change, we are consolidating all in the master config file for Windows.

Signed-off-by: Harsh Rawat <rawahars@amazon.com>