windows: add debugging option to the entrypoint script
In order to obtain code dumps on Windows, we need to set a registry key which specifies the dump folder as well as the level of dump information.

We will be adding a parameter in the entrypoint.ps1 script which the customers can enable to generate the core dumps. Since these core dump files would be generated inside the container, we need to bind mount the location to the host to retrieve the same.

Signed-off-by: Harsh Rawat <rawahars@amazon.com>