windows: moved stable version from separate file to master config
Presently, we have a separate file which stores the stable version number for Windows. This leads to the configuration being stores at multiple places. Therefore, in order to simplify the setup, we are moving the stable version in the windows.versions file.

Additionally, we have also simplified the other build scripts.

Signed-off-by: Harsh Rawat <rawahars@amazon.com>