windows: ignore windows tags while fetching latest version
Presently, we are using get_latest_dockerhub_version.py script to fetch the latest version of image present on dockerhub. Since Windows tags start with number but later have letters, the scripts are failing. To mitigate the same, we add an extra condition wherein any tag which has - in them is ignored.

Signed-off-by: Harsh Rawat <rawahars@amazon.com>