windows: decouple Linux and Windows dockerhub tags
Presently, the Python script was used to fetch the latest version on Dockerhub irrespective of the platform. This couples the Linux releases to Windows releases. In order to decouple the same, we have introduced the first param as the platform name for which we need to fetch the latest version.

Signed-off-by: Harsh Rawat <rawahars@amazon.com>